refactor: extract useSubscriptionCheckout composable, rewrite tests
- Extract checkout logic from SubscriptionRequiredDialogContentWorkspace
  into useSubscriptionCheckout composable
- Deduplicate handleAddCreditCard/handleConfirmTransition into single
  handleSubscription function
- Export pure findPlanSlug function for zero-mock unit testing
- Rewrite component tests to mock only the composable (11 slim tests)
- Add composable-level tests with minimal mocking (17 tests)
- Pure function tests for findPlanSlug with zero mocks (4 tests)

Addresses review feedback: heavy mocking signaled tight coupling.
Component is now a thin presentation layer; logic lives in the
composable where it can be tested with focused, maintainable tests.
